Buttons are fully functional and customizable and can handle any translation.
Added .::. ButtonStyle class containing onPress, onRelease, etc styles to draw the button.
Added .::. An applyStyle function that takes a string and automatically sets the action styles of the button.
Patched .::. The getMousePos and getOOP functions of Object not properly calculating relative mouse coordinates
Changed .::. The getAAP function in Object is now called getOOP. (Axis-Aligned-Point and Object-Oriented-Point)